examples/requirements.txt
pyfunctional>=1.4.3
dash-bootstrap-components>=1.2.0
dash-extensions>=0.1.4